requirements.txt
streamlit>=1.36
pandas>=2.2
python-dotenv>=1.0
requests>=2.32
plotly>=5.22
openpyxl>=3.1
altair>=5.0.0
